Moved nginx to use host port due to a weird latency issue with BGP routing
Moved to a single kube instance

I have no idea why I am getting this issue, but randomly a request will
just timeout. Like the packet is being routed to the wrong machine. Will
have to spend a lot of time debugging this if I wanted to really figure
it out.

A node server would randomly have 10-20+ second 'stalling' on a request
from the browser. Docker push would just timeout and fail to ever push
even a single layer. Moving to use hostport is a workaround for now